requirements.txt
streamlit>=1.32.0
langchain>=0.1.13
langchain-community>=0.0.29
langchain-huggingface>=0.0.1
huggingface-hub>=0.21.4
python-dotenv>=1.0.1
